requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
langchain==0.1.13
langchain-openai==0.1.1 
langchain-community==0.0.29
faiss-cpu==1.8.0.post1
tiktoken==0.6.0         
sentence-transformers==2.6.1 
python-dotenv==1.0.1     
requests==2.31.0         